+++ src/components/atoms/ModalMap/styles.js
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+const styles = color =>
+  StyleSheet.create({
+    btn: {marginTop: 10, backgroundColor: color},
+    Mapsize: {width: 300, height: 400},
+  });
+
+export default styles;
