+++ src/screens/profileScreen/index.js
@@ -6,10 +6,9 @@ import styles from './styles';
 import ProfileHeader from '../../components/atoms/ProfileHeader';
 import ProfileContent from '../../components/molecules/ProfileContent';
 
-console.log(auth().currentUser);
-const user = auth().currentUser;
-
 export default function ProfileScreen() {
+  const user = auth().currentUser;
+
   return (
     <View style={styles.background}>
       <ProfileHeader
